[oklch] Add option to display gamut-mapped color
By default, the background color of the window will use the gamut-mapped
P3 color if the current display supports P3, and the sRGB color if it
does not.  However, visualizing the difference between the gamut-mapped
P3 color and the gamut-mapped sRGB color can be very useful, so this
option allows forcing the window color to use the gamut-mapped sRGB
color even on a display that supports P3.