Add `.editorconfig` file to specify formatting
Style preferences and editor configurations vary from programmer to
programmer, but editing these files with different style settings can
result in unexpected changes in the Git history that are not related
to functionality.  The `.editorconfig` file controls (for editors that
support it), or at least documents, the expected formatting options.